move testing code to src/apps/testapps/testH3CellAreaExhaustive.c
