+++ src/apps/testapps/testGeoCoord.c
@@ -45,50 +45,6 @@ static void testDecreasingFunction(double (*function)(int),
     }
 }
 
-static void cell_area_assertions(H3Index cell) {
-    char msg[] = "cell has positive length";
-
-    t_assert(H3_EXPORT(cellAreaRads2)(cell) > 0, msg);
-    t_assert(H3_EXPORT(cellAreaKm2)(cell) > 0, msg);
-    t_assert(H3_EXPORT(cellAreaM2)(cell) > 0, msg);
-}
-
-static void earthAreaTest(int res, double (*callback)(H3Index), double target,
-                          double tol) {
-    double area = mapSumAllCells_double(res, callback);
-
-    t_assert(fabs(area - target) < tol,
-             "sum of all cells should give earth area");
-}
-
-static void commutative_distance_assertions(H3Index edge) {
-    GeoCoord a, b;
-    double ab, ba;
-
-    H3_EXPORT(h3ToGeo)
-    (H3_EXPORT(getOriginH3IndexFromUnidirectionalEdge)(edge), &a);
-    H3_EXPORT(h3ToGeo)
-    (H3_EXPORT(getDestinationH3IndexFromUnidirectionalEdge)(edge), &b);
-
-    char pos[] = "distance between cell centers should be positive";
-    char comm[] = "pairwise cell distances should be commutative";
-
-    ab = H3_EXPORT(pointDistRads)(&a, &b);
-    ba = H3_EXPORT(pointDistRads)(&b, &a);
-    t_assert(ab > 0, pos);
-    t_assert(ab == ba, comm);
-
-    ab = H3_EXPORT(pointDistKm)(&a, &b);
-    ba = H3_EXPORT(pointDistKm)(&b, &a);
-    t_assert(ab > 0, pos);
-    t_assert(ab == ba, comm);
-
-    ab = H3_EXPORT(pointDistM)(&a, &b);
-    ba = H3_EXPORT(pointDistM)(&b, &a);
-    t_assert(ab > 0, pos);
-    t_assert(ab == ba, comm);
-}
-
 SUITE(geoCoord) {
     TEST(radsToDegs) {
         double originalRads = 1;
@@ -247,50 +203,5 @@ SUITE(geoCoord) {
             t_assert(next > last, "numHexagons ordering");
             last = next;
         }
-    }
-
-    TEST(cellAreaPositive) {
-        iterateAllIndexesAtRes(0, cell_area_assertions);
-        iterateAllIndexesAtRes(1, cell_area_assertions);
-        iterateAllIndexesAtRes(2, cell_area_assertions);
-        iterateAllIndexesAtRes(3, cell_area_assertions);
-    }
-
-    TEST(cellAreaEarth) {
-        // earth area in different units
-        double rads2 = 4 * M_PI;
-        double km2 = rads2 * EARTH_RADIUS_KM * EARTH_RADIUS_KM;
-        double m2 = km2 * 1000 * 1000;
-
-        // Notice the drop in accuracy at resolution 1.
-        // I think this has something to do with Class II vs Class III
-        // resolutions.
-
-        earthAreaTest(0, H3_EXPORT(cellAreaRads2), rads2, 1e-14);
-        earthAreaTest(0, H3_EXPORT(cellAreaKm2), km2, 1e-6);
-        earthAreaTest(0, H3_EXPORT(cellAreaM2), m2, 1e0);
-
-        earthAreaTest(1, H3_EXPORT(cellAreaRads2), rads2, 1e-9);
-        earthAreaTest(1, H3_EXPORT(cellAreaKm2), km2, 1e-1);
-        earthAreaTest(1, H3_EXPORT(cellAreaM2), m2, 1e5);
-
-        earthAreaTest(2, H3_EXPORT(cellAreaRads2), rads2, 1e-12);
-        earthAreaTest(2, H3_EXPORT(cellAreaKm2), km2, 1e-5);
-        earthAreaTest(2, H3_EXPORT(cellAreaM2), m2, 1e0);
-
-        earthAreaTest(3, H3_EXPORT(cellAreaRads2), rads2, 1e-11);
-        earthAreaTest(3, H3_EXPORT(cellAreaKm2), km2, 1e-3);
-        earthAreaTest(3, H3_EXPORT(cellAreaM2), m2, 1e3);
-
-        earthAreaTest(4, H3_EXPORT(cellAreaRads2), rads2, 1e-11);
-        earthAreaTest(4, H3_EXPORT(cellAreaKm2), km2, 1e-3);
-        earthAreaTest(4, H3_EXPORT(cellAreaM2), m2, 1e2);
-    }
-
-    TEST(commutative_distances) {
-        iterateAllUnidirectionalEdgesAtRes(0, commutative_distance_assertions);
-        iterateAllUnidirectionalEdgesAtRes(1, commutative_distance_assertions);
-        iterateAllUnidirectionalEdgesAtRes(2, commutative_distance_assertions);
-        iterateAllUnidirectionalEdgesAtRes(3, commutative_distance_assertions);
     }
 }
